requirements.txt
pyjwt
bcrypt
pandas
pymongo
aiohttp
requests
reportlab
flask-mail
flask-cors
flask[async]
python-decouple
